requirements.txt
# 基本的なGoogle API関連パッケージ
google-api-python-client>=2.35.0
google-auth>=2.6.0
google-auth-oauthlib>=0.5.1
google-auth-httplib2>=0.1.0

# ファイルシステム監視用
watchdog>=2.1.7
psutil>=5.9.0

# 画像処理用
Pillow>=9.0.0

# Windows用のオプショナル依存関係（Windowsのみ必要）
pywin32>=303; sys_platform == 'win32'

# 動画再生に必要なOpenCVの依存関係
opencv-python>=4.8.0

# BGM 再生に使用
pygame>=2.5.0

# Webインターフェース用
Flask>=2.0.0
Flask-WTF>=1.0.0

# QRコード生成用
qrcode>=7.4.2
